server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlmodel>=0.0.22
aiosqlite>=0.20.0
psycopg[binary]>=3.1.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
PyJWT>=2.9.0
bcrypt>=4.2.0
